==> code/7-CNV_evolution.R <==
library(Seurat)
library(ggplot2)
library(patchwork)
library(dplyr)
library(infercnv)

younger <- readRDS("result/younger_malignant.rds")
immune <- readRDS("result/nonmalignant_subset.rds")
subset <- merge(younger, immune)

counts_matrix <- GetAssayData(object = subset, slot = 'counts')
annot <- as.data.frame(subset$tumor_ident)

# Load gene order file
gene_order <- read.table('geneLocate.txt', header = F,row.names = 1)
# Create infercnv object
infercnv_obj <- CreateInfercnvObject(raw_counts_matrix = counts_matrix,
                                     annotations_file = annot, 
                                     gene_order_file = gene_order, 
                                     ref_group_names = "non-malignant")

infercnv_obj <- infercnv::run(infercnv_obj, 
                              cutoff = 1,
                              out_dir = paste0('infercnv/younger_malignant'), 
                              cluster_by_groups=T, 
			                        resume = T,
                              analysis_mode="subclusters",
                              denoise = T, 
                              HMM = T, 
                              output_format = NA,
                              num_threads = 4)

# infercnv_obj <- readRDS("infercnv/older_malignant/run.final.infercnv_obj")
# infercnv::plot_cnv(infercnv_obj, #上两步得到的infercnv对象
#                    plot_chr_scale = T, #画染色体全长，默认只画出（分析用到的）基因
#                    output_filename = paste0('infercnv/older_malignant/infercnv.pdf'),
#                    output_format = "pdf") #保存为pdf文件
